backend/requirements.txt
fastapi
uvicorn
motor
bcrypt
python-jose
python-dotenv
email-validator
google-generativeai
python-multipart
supabase
